refactor: utils -> util 폴더 이름 변경
